+++ extensions/unifi/CHANGELOG.md
@@ -1,6 +1,14 @@
 # UniFi Changelog
 
-# [Update] - 2023-01-31
+## 2.0 - 2025-02-15
+
+- Updated to use new local UniFi API with api key
+- Added Restart device action
+- Added Device Port view
+- Added Device Radio view
+- Added live stats view
+
+## [Update] - 2023-01-31
 
 - Add support for UnifiOS 3.2+